D4 - Fixed failure with tank initialization

+++ backend/include/Thrust/EngineConfig.h
@@ -120,7 +120,7 @@ struct EngineConfig
                                int id,
                                const std::string& name,
                                const std::string& type,
-                               double tankID,
+                               int tankID,
                                double Isp,
                                double timeConstant,
                                double responseRate,
